beta 5 release candidate 7

+++ src/dg_misc.c
@@ -1,5 +1,5 @@
 /* ************************************************************************
-*   File: dg_misc.c                                       EmpireMUD 2.0b4 *
+*   File: dg_misc.c                                       EmpireMUD 2.0b5 *
 *  Usage: contains general functions for script usage.                    *
 *                                                                         *
 *  DG Scripts code had no header or author info in this file              *
@@ -27,7 +27,13 @@
 #include "db.h"
 #include "skills.h"
 
+// external funcs
+void combat_meter_damage_dealt(char_data *ch, int amt);
+void combat_meter_damage_taken(char_data *ch, int amt);
+void combat_meter_heal_dealt(char_data *ch, int amt);
+void combat_meter_heal_taken(char_data *ch, int amt);
 extern obj_data *die(char_data *ch, char_data *killer);
+extern room_data *get_room(room_data *ref, char *name);
 
 /* external vars */
 extern const char *apply_types[];
@@ -62,7 +68,6 @@ room_data *do_dg_add_room_dir(room_data *from, int dir, bld_data *bld) {
 
 	COMPLEX_DATA(new)->home_room = home;
 	COMPLEX_DATA(home)->inside_rooms++;
-	ROOM_OWNER(new) = ROOM_OWNER(home);
 	
 	if (GET_ROOM_VEHICLE(from)) {
 		++VEH_INSIDE_ROOMS(GET_ROOM_VEHICLE(from));
@@ -72,8 +77,8 @@ room_data *do_dg_add_room_dir(room_data *from, int dir, bld_data *bld) {
 		SET_BIT(ROOM_BASE_FLAGS(new), ROOM_AFF_IN_VEHICLE);
 	}
 	
-	if (ROOM_OWNER(new)) {
-		create_territory_entry(ROOM_OWNER(new), new);
+	if (ROOM_OWNER(home)) {
+		perform_claim_room(new, ROOM_OWNER(home));
 	}
 	
 	// sort now just in case
@@ -105,8 +110,12 @@ void do_dg_affect(void *go, struct script_data *sc, trig_data *trig, int script_
 	half_chop(cmd, value_p, duration_p);
 
 	/* make sure all parameters are present */
-	if (!*charname || !*property || !*value_p || !*duration_p) {
+	if (!*charname || !*property || !*value_p) {
 		script_log("Trigger: %s, VNum %d. dg_affect usage: <target> <property> <value> <duration>", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
+		return;
+	}
+	if (str_cmp(value_p, "off") && !*duration_p) {
+		script_log("Trigger: %s, VNum %d. dg_affect missing duration", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
 		return;
 	}
 
@@ -140,6 +149,10 @@ void do_dg_affect(void *go, struct script_data *sc, trig_data *trig, int script_
 		script_log("Trigger: %s, VNum %d. dg_affect: cannot locate target!", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
 		return;
 	}
+	if (IS_DEAD(ch) || EXTRACTED(ch)) {
+		// no affects on the dead
+		return;
+	}
 	
 	if (duration == -1 && !IS_NPC(ch)) {
 		script_log("Trigger: %s, VNum %d. dg_affect: cannot use infinite duration on player target", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
@@ -148,10 +161,10 @@ void do_dg_affect(void *go, struct script_data *sc, trig_data *trig, int script_
 
 	if (!str_cmp(value_p, "off")) {
 		if (type == APPLY_TYPE) {
-			affect_from_char_by_apply(ch, ATYPE_DG_AFFECT, i);
+			affect_from_char_by_apply(ch, ATYPE_DG_AFFECT, i, FALSE);
 		}
 		else {
-			affect_from_char_by_bitvector(ch, ATYPE_DG_AFFECT, BIT(i));
+			affect_from_char_by_bitvector(ch, ATYPE_DG_AFFECT, BIT(i), FALSE);
 		}
 		return;
 	}
@@ -172,6 +185,161 @@ void do_dg_affect(void *go, struct script_data *sc, trig_data *trig, int script_
 	}
 
 	affect_to_char(ch, &af);
+}
+
+
+/**
+* add/remove an affect on a room
+*/
+void do_dg_affect_room(void *go, struct script_data *sc, trig_data *trig, int script_type, char *cmd) {
+	extern const char *room_aff_bits[];
+	
+	char junk[MAX_INPUT_LENGTH]; /* will be set to "dg_affect_room" */
+	char roomname[MAX_INPUT_LENGTH], property[MAX_INPUT_LENGTH];
+	char value_p[MAX_INPUT_LENGTH], duration_p[MAX_INPUT_LENGTH];
+	bitvector_t i = 0, type = 0;
+	struct affected_type af;
+	room_data *room = NULL;
+	int duration = 0;
+
+	half_chop(cmd, junk, cmd);
+	half_chop(cmd, roomname, cmd);
+	half_chop(cmd, property, cmd);
+	half_chop(cmd, value_p, duration_p);
+
+	/* make sure all parameters are present */
+	if (!*roomname || !*property || !*value_p || !*duration_p) {
+		script_log("Trigger: %s, VNum %d. dg_affect_room usage: <room> <property> <on|off> <duration>", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
+		return;
+	}
+	
+	duration = atoi(duration_p);
+	if (duration == 0 || duration < -1) {
+		script_log("Trigger: %s, VNum %d. dg_affect_room: need positive duration!", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
+		script_log("Line was: dg_affect_room %s %s %s %s (%d)", roomname, property, value_p, duration_p, duration);
+		return;
+	}
+
+	// find the property -- search room_aff_bits
+	if ((i = search_block(property, room_aff_bits, TRUE)) != NOTHING) {
+		type = AFFECT_TYPE;
+	}
+
+	if (!type) { // property not found
+		script_log("Trigger: %s, VNum %d. dg_affect_room: unknown property '%s'!", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig), property);
+		return;
+	}
+
+	/* locate the target */
+	room = get_room((script_type == WLD_TRIGGER ? (room_data*)go : NULL), roomname);
+	if (!room) {
+		script_log("Trigger: %s, VNum %d. dg_affect_room: cannot locate target: %s", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig), roomname);
+		return;
+	}
+	
+	if (duration == -1 && IS_OUTDOOR_TILE(room) && GET_ROOM_VNUM(room) < MAP_SIZE) {
+		script_log("Trigger: %s, VNum %d. dg_affect_room: cannot use infinite duration on map tile target", GET_TRIG_NAME(trig), GET_TRIG_VNUM(trig));
+		return;
+	}
+
+	if (!str_cmp(value_p, "off")) {
+		affect_from_room_by_bitvector(room, ATYPE_DG_AFFECT, BIT(i), FALSE);
+		return;
+	}
+
+	/* add the affect */
+	af.type = ATYPE_DG_AFFECT;
+	af.cast_by = (script_type == MOB_TRIGGER ? CAST_BY_ID((char_data*)go) : 0);
+	af.duration = (duration == -1 ? UNLIMITED : ceil((double)duration / SECS_PER_MUD_HOUR));
+	af.modifier = 0;
+	af.location = APPLY_NONE;
+	af.bitvector = BIT(i);
+
+	affect_to_room(room, &af);
+}
+
+
+/**
+* Do the actual work for the %build% commands, once everything has been
+* validated except the last arg.
+*
+* @param room_data *target The room to change.
+* @param char *argument <vnum [dir] | ruin | demolish>
+*/
+void do_dg_build(room_data *target, char *argument) {
+	void complete_building(room_data *room);
+	void ruin_one_building(room_data *room);	// db.world.c
+	void special_building_setup(char_data *ch, room_data *room);
+	extern const int rev_dir[];
+	
+	char vnum_arg[MAX_INPUT_LENGTH], dir_arg[MAX_INPUT_LENGTH];
+	bool ruin = FALSE, demolish = FALSE;
+	any_vnum vnum = NOTHING;
+	bld_data *bld = NULL;
+	int dir = NORTH;
+	
+	skip_spaces(&argument);
+	
+	if (!target || !*argument) {
+		return;
+	}
+	
+	// parse arg
+	if (is_abbrev(argument, "ruin")) {
+		ruin = TRUE;
+	}
+	else if (is_abbrev(argument, "demolish")) {
+		demolish = TRUE;
+	}
+	else if (isdigit(*argument)) {
+		argument = any_one_arg(argument, vnum_arg);
+		argument = any_one_arg(argument, dir_arg);
+		
+		if ((vnum = atoi(vnum_arg)) < 0 || !(bld = building_proto(vnum)) || IS_SET(GET_BLD_FLAGS(bld), BLD_ROOM)) {
+			script_log("do_dg_build: invalid building: %d", vnum);
+			return;
+		}
+		else if (*dir_arg && ((dir = parse_direction(NULL, dir_arg)) == NO_DIR || dir >= NUM_2D_DIRS)) {
+			script_log("do_dg_build: invalid direction: %s", dir_arg);
+			return;
+		}
+		else if (dir != NO_DIR && (!SHIFT_DIR(target, dir) || (IS_SET(GET_BLD_FLAGS(bld), BLD_TWO_ENTRANCES) && !SHIFT_DIR(target, rev_dir[dir])))) {
+			script_log("do_dg_build: invalid direction from room %d: %s", GET_ROOM_VNUM(target), dir_arg);
+			return;
+		}
+		
+		// bld validated
+	}
+	else {
+		script_log("do_dg_build: invalid argument: %s", argument);
+		return;
+	}
+	
+	
+	if (ruin) {
+		room_data *home = HOME_ROOM(target);
+		if (GET_ROOM_VNUM(home) < MAP_SIZE && GET_BUILDING(home)) {
+			ruin_one_building(home);
+		}
+	}
+	else if (demolish) {
+		disassociate_building(target);
+	}
+	else if (bld) {
+		disassociate_building(target);
+	
+		construct_building(target, GET_BLD_VNUM(bld));
+		special_building_setup(NULL, target);
+		complete_building(target);
+	
+		if (dir != NO_DIR) {
+			create_exit(target, SHIFT_DIR(target, dir), dir, FALSE);
+			if (IS_SET(GET_BLD_FLAGS(bld), BLD_TWO_ENTRANCES)) {
+				create_exit(target, SHIFT_DIR(target, rev_dir[dir]), rev_dir[dir], FALSE);
+			}
+			COMPLEX_DATA(target)->entrance = rev_dir[dir];
+		}
+	}
 }
 
 
@@ -226,6 +394,57 @@ void do_dg_own(empire_data *emp, char_data *vict, obj_data *obj, room_data *room
 
 
 /**
+* Purges all matching vnums in an instance. This is called from:
+* %purge% instance <arguments>
+*
+* Currently it supports:
+* - mob <vnum> [message]
+*
+* @param void *owner The owner of the script, to check if it purges itself.
+* @param struct instance_data *inst The instance whose contents are being purged.
+* @param char *argument The arguments passed to "%purge% instance".
+*/
+void dg_purge_instance(void *owner, struct instance_data *inst, char *argument) {
+	char arg1[MAX_INPUT_LENGTH], arg2[MAX_INPUT_LENGTH];
+	char_data *mob, *next_mob;
+	any_vnum vnum;
+	
+	if (!inst) {
+		return;
+	}
+	
+	argument = any_one_arg(argument, arg1);	// type
+	argument = any_one_arg(argument, arg2);	// vnum
+	skip_spaces(&argument);	// message
+	vnum = atoi(arg2);
+	
+	if (!*arg1 || !*arg2) {
+		script_log("dg_purge_instance called with invalid arguments: %s %s %s", arg1, arg2, argument);
+	}
+	else if (is_abbrev(arg1, "mobile")) {
+		LL_FOREACH_SAFE(character_list, mob, next_mob) {
+			if (!IS_NPC(mob) || GET_MOB_VNUM(mob) != vnum || EXTRACTED(mob)) {
+				continue;
+			}
+			
+			// found
+			if (*argument) {
+				act(argument, TRUE, mob, NULL, NULL, TO_ROOM);
+			}
+			
+			if (mob == owner) {
+				dg_owner_purged = 1;
+			}
+			extract_char(mob);
+		}
+	}
+	else {
+		script_log("dg_purge_instance called with invalid type: %s", arg1);
+	}
+}
+
+
+/**
 * Processes the %quest% command.
 *
 * @param int go_type _TRIGGER type for 'go'
@@ -255,6 +474,10 @@ void do_dg_quest(int go_type, void *go, char *argument) {
 	}
 	if (!isdigit(*vnum_arg) || (vnum = atoi(vnum_arg)) < 0 || !(quest = quest_proto(vnum))) {
 		script_log_by_type(go_type, go, "dg_quest: invalid vnum '%s'", vnum_arg);
+		return;
+	}
+	if (QUEST_FLAGGED(quest, QST_IN_DEVELOPMENT)) {
+		script_log_by_type(go_type, go, "dg_quest: quest [%d] %s is set IN-DEVELOPMENT", QUEST_VNUM(quest), QUEST_NAME(quest));
 		return;
 	}
 	
@@ -345,6 +568,9 @@ void do_dg_quest(int go_type, void *go, char *argument) {
 	else if (is_abbrev(cmd_arg, "trigger")) {
 		qt_triggered_task(vict, QUEST_VNUM(quest));
 	}
+	else if (is_abbrev(cmd_arg, "untrigger")) {
+		qt_untrigger_task(vict, QUEST_VNUM(quest));
+	}
 	else {
 		script_log_by_type(go_type, go, "dg_quest: invalid command '%s'", cmd_arg);
 	}
@@ -360,13 +586,10 @@ void do_dg_quest(int go_type, void *go, char *argument) {
 */
 void do_dg_terracrop(room_data *target, crop_data *cp) {
 	sector_data *sect;
-	empire_data *emp;
 	
 	if (!target || !cp) {
 		return;
 	}
-	
-	emp = ROOM_OWNER(target);
 	
 	if (!(sect = find_first_matching_sector(SECTF_CROP, NOBITS))) {
 		// no crop sects?
@@ -378,10 +601,18 @@ void do_dg_terracrop(room_data *target, crop_data *cp) {
 		
 		remove_depletion(target, DPLTN_PICK);
 		remove_depletion(target, DPLTN_FORAGE);
-	}
-	
-	if (emp) {
-		read_empire_territory(emp);
+		remove_depletion(target, DPLTN_DIG);
+		remove_depletion(target, DPLTN_GATHER);
+		remove_depletion(target, DPLTN_FISH);
+		remove_depletion(target, DPLTN_QUARRY);
+		remove_depletion(target, DPLTN_PAN);
+		remove_depletion(target, DPLTN_TRAPPING);
+		remove_depletion(target, DPLTN_CHOP);
+		
+		if (ROOM_OWNER(target)) {
+			void deactivate_workforce_room(empire_data *emp, room_data *room);
+			deactivate_workforce_room(ROOM_OWNER(target), target);
+		}
 	}
 }
 
@@ -395,14 +626,12 @@ void do_dg_terracrop(room_data *target, crop_data *cp) {
 */
 void do_dg_terraform(room_data *target, sector_data *sect) {
 	sector_data *old_sect;
-	empire_data *emp;
 	
 	if (!target || !sect) {
 		return;
 	}
 	
 	old_sect = BASE_SECT(target);
-	emp = ROOM_OWNER(target);
 	
 	change_terrain(target, GET_SECT_VNUM(sect));
 	
@@ -410,9 +639,20 @@ void do_dg_terraform(room_data *target, sector_data *sect) {
 	if (IS_ROAD(target)) {
 		change_base_sector(target, old_sect);
 	}
-
-	if (emp) {
-		read_empire_territory(emp);
+	
+	remove_depletion(target, DPLTN_PICK);
+	remove_depletion(target, DPLTN_FORAGE);
+	remove_depletion(target, DPLTN_DIG);
+	remove_depletion(target, DPLTN_GATHER);
+	remove_depletion(target, DPLTN_FISH);
+	remove_depletion(target, DPLTN_QUARRY);
+	remove_depletion(target, DPLTN_PAN);
+	remove_depletion(target, DPLTN_TRAPPING);
+	remove_depletion(target, DPLTN_CHOP);
+	
+	if (ROOM_OWNER(target)) {
+		void deactivate_workforce_room(empire_data *emp, room_data *room);
+		deactivate_workforce_room(ROOM_OWNER(target), target);
 	}
 }
 
@@ -469,6 +709,35 @@ int valid_dg_target(char_data *ch, int bitvector) {
 
 
 /**
+* Runs triggers to update and prepare the mud at startup.
+*/
+void run_reboot_triggers(void) {
+	void reboot_mtrigger(char_data *ch);
+	void reboot_otrigger(obj_data *obj);
+	void reboot_vtrigger(vehicle_data *veh);
+	void reboot_wtrigger(room_data *room);
+	
+	vehicle_data *veh, *next_veh;
+	room_data *room, *next_room;
+	char_data *mob, *next_mob;
+	obj_data *obj, *next_obj;
+	
+	HASH_ITER(hh, world_table, room, next_room) {
+		reboot_wtrigger(room);
+	}
+	LL_FOREACH_SAFE(character_list, mob, next_mob) {
+		reboot_mtrigger(mob);
+	}
+	LL_FOREACH_SAFE(object_list, obj, next_obj) {
+		reboot_otrigger(obj);
+	}
+	LL_FOREACH_SAFE(vehicle_list, veh, next_veh) {
+		reboot_vtrigger(veh);
+	}
+}
+
+
+/**
 * Deals damage to a character based on scaled level and modifier.
 *
 * @param char_data *vict The poor sod who's taking damage.
@@ -481,6 +750,8 @@ void script_damage(char_data *vict, char_data *killer, int level, int dam_type, 
 	void death_log(char_data *ch, char_data *killer, int type);
 	extern char *get_room_name(room_data *room, bool color);
 	extern int reduce_damage_from_skills(int dam, char_data *victim, char_data *attacker, int damtype);
+	void scale_mob_for_character(char_data *mob, char_data *ch);
+	void scale_mob_to_level(char_data *mob, int level);
 	
 	double dam;
 	
@@ -488,23 +759,49 @@ void script_damage(char_data *vict, char_data *killer, int level, int dam_type, 
 	if (IS_DEAD(vict)) {
 		return;
 	}
+	if (AFF_FLAGGED(vict, AFF_NO_ATTACK)) {
+		return;	// can't be attacked
+	}
 	
 	if (IS_IMMORTAL(vict) && (modifier > 0)) {
 		msg_to_char(vict, "Being the cool immortal you are, you sidestep a trap, obviously placed to kill you.\r\n");
 		return;
 	}
 	
-	dam = level / 8.0;
+	// check scaling
+	if (killer && IS_NPC(killer) && GET_CURRENT_SCALE_LEVEL(killer) == 0) {
+		scale_mob_for_character(killer, vict);
+	}
+	if (IS_NPC(vict) && GET_CURRENT_SCALE_LEVEL(vict) == 0) {
+		if (killer) {
+			scale_mob_for_character(vict, killer);
+		}
+		else {
+			scale_mob_to_level(vict, level);
+		}
+	}
+	
+	dam = level / 7.0;
 	dam *= modifier;
 	
 	// guarantee at least 1
 	if (modifier > 0) {
 		dam = reduce_damage_from_skills(dam, vict, killer, dam_type);	// resistance, etc
 		dam = MAX(1, dam);
+		
+		if (killer && vict != killer) {
+			combat_meter_damage_dealt(killer, dam);
+		}
+		combat_meter_damage_taken(vict, dam);
 	}
 	else if (modifier < 0) {
 		// healing
 		dam = MIN(-1, dam);
+		
+		if (killer) {
+			combat_meter_heal_dealt(killer, -dam);
+		}
+		combat_meter_heal_taken(vict, -dam);
 	}
 	
 	GET_HEALTH(vict) -= dam;
@@ -520,7 +817,11 @@ void script_damage(char_data *vict, char_data *killer, int level, int dam_type, 
 		}
 		die(vict, killer ? killer : vict);
 	}
-}  
+	else if (modifier < 0 && GET_HEALTH(vict) > 0 && GET_POS(vict) < POS_SLEEPING) {
+		msg_to_char(vict, "You recover and wake up.\r\n");
+		GET_POS(vict) = IS_NPC(vict) ? POS_STANDING : POS_SITTING;
+	}
+}
 
 
 /**
@@ -547,7 +848,7 @@ void script_damage_over_time(char_data *vict, int level, int dam_type, double mo
 		return;
 	}
 	
-	dam = level / 20.0;
+	dam = level / 17.5;
 	dam *= modifier;
 	
 	// guarantee at least 1
